refactor: adapt /prediction/ to use versions
